Split mod and improve partitioning algo
